requirements.txt
Flask >= 2.3.0
pandas >= 2.0.2
matplotlib >= 3.7.0
numpy >= 1.24.3
Pillow >= 9.5.0
protobuf <= 3.20.1

